wip: code seems good, but wierdness on value sets save, some values disappearing
