+++ frontend-react/src/hooks/UseValueSets.ts
@@ -1,17 +1,13 @@
 import { useCallback } from "react";
-import { AxiosRequestConfig } from "axios";
 import { useQuery, useMutation } from "@tanstack/react-query";
-import omit from "lodash.omit";
 
+import { useAuthorizedFetch } from "../contexts/AuthorizedFetchContext";
 import {
+    lookupTablesEndpoints,
     LookupTable,
     ValueSet,
     ValueSetRow,
-} from "../network/api/LookupTableApi";
-import { useAuthorizedFetch } from "../contexts/AuthorizedFetchContext";
-import { StringIndexed } from "../utils/UsefulTypes";
-
-import { EndpointConfig, HTTPMethods } from "./UseCreateFetch";
+} from "../config/endpoints/lookupTables";
 
 export interface TableAttributes {
     version: number;
@@ -19,68 +15,8 @@ export interface TableAttributes {
     createdBy?: string;
 }
 
-// these can be stored on our resource once we build that out
-// notice the react-router style colon demarcated dynamic path segments
-const getTableListConfig = {
-    path: "/lookuptables/list",
-    method: HTTPMethods.GET,
-};
-const getTableDataConfig = {
-    path: "/lookuptables/:tableName/:version/content",
-    method: HTTPMethods.GET,
-};
-
-const updateTableConfig = {
-    path: "/lookuptables/:tableName",
-    method: HTTPMethods.POST,
-};
-const activateTableConfig = {
-    path: "/lookuptables/:tableName/:version/activate",
-    method: HTTPMethods.PUT,
-};
-
-interface AxiosOptionsWithSegments extends AxiosRequestConfig {
-    segments: StringIndexed<string>;
-}
-
-// takes a map of path segment keys (as defined by colons in config paths declarations, such as in react outer)
-// to segment values for a particular API call
-// in order to produce concrete path from a dynamic one
-// this stuff likely will live on even in a world with resources, one way or another
-const hydrateDynamicPathSegments = (
-    path: string,
-    segments?: StringIndexed<string>
-) => {
-    if (!segments) {
-        return path;
-    }
-    const pathWithSegments = Object.entries(segments).reduce(
-        (pathWithSegments, [segmentKey, segmentValue]) =>
-            pathWithSegments.replace(`:${segmentKey}`, segmentValue),
-        path
-    );
-    if (pathWithSegments.indexOf("/:") > -1) {
-        throw new Error(`missing dynamic path param: ${path}, ${segments}`);
-    }
-    return pathWithSegments;
-};
-
-// this could live on each resource, which would preclude the need to pass in the config, as that
-// would also be stored on the resource
-const toFetchParams = (
-    baseConfig: EndpointConfig,
-    requestParams: Partial<AxiosOptionsWithSegments>
-) => {
-    const pathWithSegments = hydrateDynamicPathSegments(
-        baseConfig.path,
-        requestParams.segments
-    );
-    return {
-        method: baseConfig.method,
-        path: pathWithSegments,
-        options: { ...omit(requestParams, "segments") }, // this is yucky
-    };
-};
+const { getTableData, getTableList, updateTable, activateTable } =
+    lookupTablesEndpoints;
 
 /*
 
@@ -144,8 +80,8 @@ export const useValueSetsTable = <T extends ValueSet | ValueSetRow>(
 
     // get all lookup tables
     const { error: tableError, data: tableData } = useQuery<LookupTable[]>(
-        ["lookupTables"],
-        () => lookupTableFetch(getTableListConfig),
+        [getTableList.queryKey],
+        () => lookupTableFetch(getTableList),
         { enabled: !suppliedVersion } // only if version was not already passed in
     );
 
@@ -168,14 +104,12 @@ export const useValueSetsTable = <T extends ValueSet | ValueSetRow>(
     // create the function to use for fetching table data from the API
     const memoizedDataFetch = useCallback(
         () =>
-            dataFetch(
-                toFetchParams(getTableDataConfig, {
-                    segments: {
-                        tableName: dataTableName,
-                        version: `${versionData!!.version!}`, // number to string,
-                    },
-                })
-            ),
+            dataFetch(getTableData, {
+                segments: {
+                    tableName: dataTableName,
+                    version: `${versionData!!.version!}`, // number to string,
+                },
+            }),
         [dataFetch, versionData, dataTableName]
     );
 
@@ -183,7 +117,7 @@ export const useValueSetsTable = <T extends ValueSet | ValueSetRow>(
     // does not seem entirely worth it to add the complexity needed to account for that on the frontend, better
     // to make the API conform better to the frontend's expectations. TODO: look at this when refactoring the API
     const { error: valueSetError, data: valueSetData } = useQuery<Array<T>>(
-        ["lookupTable", versionData?.version, dataTableName],
+        [getTableData.queryKey, versionData?.version, dataTableName],
         memoizedDataFetch,
         { enabled: !!versionData?.version }
     );
@@ -219,12 +153,10 @@ export const useValueSetUpdate = () => {
     const valueSetFetch = useAuthorizedFetch<LookupTable>();
 
     const updateValueSet = ({ data, tableName }: UpdateValueSetOptions) => {
-        return valueSetFetch(
-            toFetchParams(updateTableConfig, {
-                segments: { tableName: tableName },
-                data,
-            })
-        );
+        return valueSetFetch(updateTable, {
+            segments: { tableName: tableName },
+            data,
+        });
     };
 
     // generic signature is defined here https://github.com/TanStack/query/blob/4690b585722d2b71d9b87a81cb139062d3e05c9c/packages/react-query/src/useMutation.ts#L66
@@ -245,14 +177,12 @@ export const useValueSetActivation = () => {
         tableVersion,
         tableName,
     }: ActivateValueSetOptions) => {
-        return valueSetFetch(
-            toFetchParams(activateTableConfig, {
-                segments: {
-                    tableName,
-                    version: `${tableVersion}`,
-                },
-            })
-        );
+        return valueSetFetch(activateTable, {
+            segments: {
+                tableName,
+                version: `${tableVersion}`,
+            },
+        });
     };
     const mutation = useMutation<LookupTable, Error, ActivateValueSetOptions>(
         activateValueSet
